add free form text search
Also, moved JobSearchItem into its own file at long last
